refactor(appshell): extract 5 data-layer hooks + stabilize conversation re-renders
Phase 1 split: extract useWorkspaceToast, useResolvedShortcuts, useSessionRunStates, useNavigationSidebar, useSettingsOpenHandlers from the ~1650-line AppShell. App.tsx 1976 to 1784 lines, behavior verbatim.

Perf (P0/P1): useMemo-cache sessionSelectionHistory + FileLinkProvider value + header JSX nodes, which were busting WorkspaceConversationContainer's memo on unrelated AppShell renders (React Compiler bailed out on the oversized component).

Verified: tsc + biome + 1319 frontend tests green, zero regression.
